Improved the behaviour of ARG_SEARCH_STRICT.
This commit also fixes a huge bug in the argument-processor, in which
the `options` value has not large enough. A small comment has been added
to explain the importance.

+++ args.h
@@ -49,7 +49,8 @@ enum arg_positions_t {
         ARG_COLOUR_OUTPUT    = 4096  /* 13th bit */
 };
 
-typedef uint8_t opts_t;
+/* uintN_t, such that log2(<highest `arg_position_t`>) <= N */
+typedef uint16_t opts_t;
 
 extern opts_t  options;
 int process_args ( int, char **, int * );
